Major cleanup of tally filter interfaces

+++ include/openmc/capi.h
@@ -17,8 +17,8 @@ extern "C" {
   int openmc_cell_set_fill(int32_t index, int type, int32_t n, const int32_t* indices);
   int openmc_cell_set_id(int32_t index, int32_t id);
   int openmc_cell_set_temperature(int32_t index, double T, const int32_t* instance);
-  int openmc_energy_filter_get_bins(int32_t index, double** energies, int32_t* n);
-  int openmc_energy_filter_set_bins(int32_t index, int32_t n, const double* energies);
+  int openmc_energy_filter_get_bins(int32_t index, double** energies, size_t* n);
+  int openmc_energy_filter_set_bins(int32_t index, size_t n, const double* energies);
   int openmc_extend_cells(int32_t n, int32_t* index_start, int32_t* index_end);
   int openmc_extend_filters(int32_t n, int32_t* index_start, int32_t* index_end);
   int openmc_extend_materials(int32_t n, int32_t* index_start, int32_t* index_end);
@@ -57,8 +57,8 @@ extern "C" {
   int openmc_material_set_densities(int32_t index, int n, const char** name, const double* density);
   int openmc_material_set_id(int32_t index, int32_t id);
   int openmc_material_set_volume(int32_t index, double volume);
-  int openmc_material_filter_get_bins(int32_t index, int32_t** bins, int32_t* n);
-  int openmc_material_filter_set_bins(int32_t index, int32_t n, const int32_t* bins);
+  int openmc_material_filter_get_bins(int32_t index, int32_t** bins, size_t* n);
+  int openmc_material_filter_set_bins(int32_t index, size_t n, const int32_t* bins);
   int openmc_mesh_filter_get_mesh(int32_t index, int32_t* index_mesh);
   int openmc_mesh_filter_set_mesh(int32_t index, int32_t index_mesh);
   int openmc_mesh_get_id(int32_t index, int32_t* id);
